+++ op_edgepatches/edgepatches_ui.py
@@ -33,8 +33,9 @@ import os
 import copy
 
 from ..lib import common_utilities
-from ..lib.common_utilities import bversion, get_object_length_scale, dprint, profiler, frange, selection_mouse, showErrorMessage
+from ..lib.common_utilities import bversion, get_object_length_scale, dprint, frange, selection_mouse, showErrorMessage
 from ..lib.common_utilities import point_inside_loop2d, get_source_object
+from ..lib.classes.profiler.profiler import profiler
 from ..lib.classes.textbox.textbox import TextBox
 from ..lib.classes.sketchbrush.sketchbrush import SketchBrush
 from .. import key_maps
